Update EasyAnimateV2 (#5)
* update EasyAnimateV2

* update datasets loader

* update datasets loader

* update fast api

* complete data preprocess pipeline.

* update lots of readme

* update readme bans

* fix bug in validation while training

* provide example for video cut

* update text box

* add arxiv

* delete IDDPM

* update gallery

* update arxiv

* update readme

* update readme

* link fix

* update vae readme

---------

+++ predict_t2i.py
@@ -25,7 +25,7 @@ lora_path           = None
 
 # Other params
 sample_size     = [512, 512]
-weight_dtype    = torch.float16
+weight_dtype    = torch.bfloat16
 prompt          = "1girl, bangs, blue eyes, blunt bangs, blurry, blurry background, bob cut, depth of field, lips, looking at viewer, motion blur, nose, realistic, red lips, shirt, short hair, solo, white shirt."
 negative_prompt = "bad detailed"
 guidance_scale  = 6.0
